+++ src/main/java/net/tassia/pancake/plugin/core/CorePlugin.kt
@@ -1,8 +1,16 @@
 package net.tassia.pancake.plugin.core
 
+import net.tassia.event.Event
 import net.tassia.pancake.Pancake
 import net.tassia.pancake.plugin.Plugin
 import net.tassia.pancake.plugin.PluginInfo
+import net.tassia.pancake.plugin.core.event.mail.IncomingMailEvent
+import net.tassia.pancake.plugin.core.event.mail.MailRouteEvent
+import net.tassia.pancake.plugin.core.event.mail.MailRoutedEvent
+import net.tassia.pancake.plugin.core.listener.mail.CoreIncomingMailListener
+import net.tassia.pancake.plugin.core.listener.mail.CoreMailRouteListener
+import net.tassia.pancake.plugin.core.listener.mail.CoreMailRoutedListener
+import kotlin.reflect.KClass
 
 /**
  * The core plugin for Pancake. Adds the base functions.
@@ -16,10 +24,17 @@ class CorePlugin(override val pancake: Pancake) : Plugin(pancake) {
 
 	override val info: PluginInfo = Info
 
+	override val events: Set<KClass<out Event>> = setOf(
+		IncomingMailEvent::class, MailRoutedEvent::class, MailRouteEvent::class
+	)
+
 
 
 	override fun onEnable() {
-		// TODO
+		// Register listeners
+		pancake.events.registerListener(CoreIncomingMailListener)
+		pancake.events.registerListener(CoreMailRoutedListener)
+		pancake.events.registerListener(CoreMailRouteListener)
 	}
