+++ src/main/java/net/tassia/pancake/plugin/core/event/mail/MailRouteEvent.kt
@@ -1,4 +1,4 @@
-package net.tassia.pancake.event
+package net.tassia.pancake.plugin.core.event.mail
 
 import net.tassia.event.Cancellable
 import net.tassia.pancake.Pancake
@@ -6,6 +6,8 @@ import net.tassia.pancake.entity.account.Account
 import net.tassia.pancake.entity.folder.Folder
 import net.tassia.pancake.entity.mail.Mail
 import net.tassia.pancake.entity.route.Route
+import net.tassia.pancake.event.MailEvent
+import net.tassia.pancake.event.PancakeEvent
 
 /**
  * Called when a [Mail] is about to be routed.
@@ -15,12 +17,8 @@ import net.tassia.pancake.entity.route.Route
  */
 data class MailRouteEvent(
 
+	override val mail: Mail,
 	override val pancake: Pancake,
-
-	/**
-	 * The mail that should be routed.
-	 */
-	val mail: Mail,
 
 	/**
 	 * The receiver address to use for routing.
@@ -44,4 +42,4 @@ data class MailRouteEvent(
 
 	override var isCancelled: Boolean = false
 
-) : PancakeEvent(pancake), Cancellable
+) : MailEvent(mail, pancake), Cancellable
