+++ src/main/java/net/tassia/pancake/plugin/core/listener/mail/CoreIncomingMailListener.kt
@@ -1,9 +1,9 @@
-package net.tassia.pancake.listener
+package net.tassia.pancake.plugin.core.listener.mail
 
 import net.tassia.event.EventListener
-import net.tassia.pancake.event.IncomingMailEvent
-import net.tassia.pancake.event.MailRouteEvent
-import net.tassia.pancake.event.MailRoutedEvent
+import net.tassia.pancake.plugin.core.event.mail.IncomingMailEvent
+import net.tassia.pancake.plugin.core.event.mail.MailRouteEvent
+import net.tassia.pancake.plugin.core.event.mail.MailRoutedEvent
 
 /**
  * Listens for incoming mails and does magic do make sure it is treated right.
@@ -24,12 +24,12 @@ object CoreIncomingMailListener : EventListener<IncomingMailEvent> {
 
 		// Tell everyone that we want to route this mail
 		pan.logger.info("Mail ${event.mail} received. Routing it for $address")
-		val routeEvent = MailRouteEvent(pan, event.mail, address, null, null, null)
+		val routeEvent = MailRouteEvent(event.mail, pan, address, null, null, null)
 		pan.events.callEvent(routeEvent)
 
 		// Tell everyone that we routed this mail
 		pan.logger.info("Mail ${event.mail} received and routed for $address")
-		pan.events.callEvent(MailRoutedEvent(pan, event.mail, address, routeEvent.account, routeEvent.folder, routeEvent.route))
+		pan.events.callEvent(MailRoutedEvent(event.mail, pan, address, routeEvent.account, routeEvent.folder, routeEvent.route))
 	}
 
 }
